scenario-generators/pyproject.toml

[tool.poetry]
name = "gitops-scenarios"
version = "0.1.0"
description = "A collection of scripts for generating k8s resources to test weave-gitops with"
authors = ["weaveworks"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
